cambiato da double a float attributo creditoofferto nell OffertaModel

+++ Backend/ProgettoAste2/src/main/java/com/gruppoaste2/progettoaste/model/OffertaModel.java
@@ -4,17 +4,17 @@ import java.util.Date;
 
 public class OffertaModel {
 
-    private final double creditoOfferto;
+    private final float creditoOfferto;
     private final Date dataOfferta;
     private final UtenteRegistratoModel utenteOfferente;
 
-    public OffertaModel(double creditoOfferto, Date dataOfferta, UtenteRegistratoModel utenteOfferente) {
+    public OffertaModel(float creditoOfferto, Date dataOfferta, UtenteRegistratoModel utenteOfferente) {
         this.creditoOfferto = creditoOfferto;
         this.dataOfferta = dataOfferta;
         this.utenteOfferente = utenteOfferente;
     }
 
-    public double getCreditoOfferto() {
+    public float getCreditoOfferto() {
         return creditoOfferto;
     }
